requirements.txt
python-docx
spacy
docx2pdf
fpdf2
